tasks core: create a common core for Task and Option Mixin classes
The idea here is to help simplify the API and the amount of boilerplate
required to make all the pieces work together while still maintaining a
relatively easy to maintain set of tasks.

Config option loading for tasks is now delayed until the last minute
rather than on initialization. This allows for more automatic
cooperation among the inheritence tree.