tasks: increase default verbosity of tasks for better feedback
Some commands that tasks execute are left to print their own stdout to
the stdout of escadrille. So to make the default output make more sense
escadrille tasks should print something indicating their start and the
escadrille engine will print the result. This also effectively bookmarks
and identifies the output of sub-commands that the tasks execute.